+++ src/Blocks/components/form/assets/utils.js
@@ -911,7 +911,7 @@ export class Utils {
 		setStateValues(name, value, formId);
 		this.setFieldFilledState(formId, name);
 
-		this.enrichment.setLocalStorageFormPrefillItem(formId, name);
+		this.enrichment.setLocalStorageFormPrefillField(formId, name);
 
 		this.conditionalTags.setField(formId, name);
 	}
@@ -950,7 +950,7 @@ export class Utils {
 
 		setStateValues(name, value, formId);
 
-		this.enrichment.setLocalStorageFormPrefillItem(formId, name);
+		this.enrichment.setLocalStorageFormPrefillField(formId, name);
 
 		this.conditionalTags.setField(formId, name);
 
@@ -999,7 +999,7 @@ export class Utils {
 
 		this.setFieldFilledState(formId, name);
 
-		this.enrichment.setLocalStorageFormPrefillItem(formId, name);
+		this.enrichment.setLocalStorageFormPrefillField(formId, name);
 
 		this.conditionalTags.setField(formId, name);
 	}
@@ -1064,7 +1064,7 @@ export class Utils {
 		setStateValues(name, { ...newValue, ...value }, formId);
 		this.setFieldFilledState(formId, name);
 
-		this.enrichment.setLocalStorageFormPrefillItem(formId, name);
+		this.enrichment.setLocalStorageFormPrefillField(formId, name);
 
 		this.conditionalTags.setField(formId, name);
 	}
@@ -1126,7 +1126,7 @@ export class Utils {
 			}
 		}
 
-		this.enrichment.setLocalStorageFormPrefillItem(formId, name);
+		this.enrichment.setLocalStorageFormPrefillField(formId, name);
 
 		this.conditionalTags.setField(formId, name);
 	}
@@ -1175,7 +1175,7 @@ export class Utils {
 			}
 		}
 
-		this.enrichment.setLocalStorageFormPrefillItem(formId, name);
+		this.enrichment.setLocalStorageFormPrefillField(formId, name);
 
 		this.conditionalTags.setField(formId, name);